Added some home page text and a button
Things are starting to shape up pretty well around here... we might
have ourselves a website soon enough!  Next goal is to add another cool
content section to the home page that links to the g3 article, then I
want to flesh out the past articles page... I have some loose ideas for
what that one is going to look like, but for now it's all just in my
head!