Finished the contact page
The contact page is now done and "Working", thought the form submission
really doesn't do anything, lol.
Overall I'm like, decently pleased with the visuals for the site, but
not super stoked.  I spent the majority of my time, by far, on the
featured article's page, which may or may not have been a good choice.

Hopefully whoever's grading this site will check out some of the past
articles, because a couple of them (*coug* TBH5 *cough*) are pretty
darn schnazzy